+++ src/z-file.c
@@ -918,10 +918,11 @@ void file_get_savefile(char *filename, size_t max, const char *base,
 		strnfmt(filename, max, "%s/temp%u.%s", dirname(base), count,
 			ext);
 #else
-	strnfmt(filename, max, "%s%u.%s", base, Rand_simple(1000000), ext);
+	strnfmt(filename, max, "%s%lu.%s",
+		base, (unsigned long)Rand_simple(1000000), ext);
 	while (file_exists(filename) && (count++ < 100))
-		strnfmt(filename, max, "%s%u%u.%s", base,
-			Rand_simple(1000000), count, ext);
+		strnfmt(filename, max, "%s%lu%u.%s",
+			base, (unsigned long)Rand_simple(1000000), count, ext);
 #endif /* ! DJGPP */
 	return;
 }
